Extract the declassification authority surface from kernel.rs (line ratchet)
The ratchet did its job: kernel.rs may not grow (ceiling 2479, the one-shot
work took it to 2516). set_trusted_keys + apply_declassification_token move
to kernel/declassify_authority.rs — a CHILD module, so the impl reaches the
kernel's private fields without widening trusted_public_keys or the spent
set's visibility. kernel.rs: 2516 → 2430. No behavior change; kernel_token
9/9.
